+++ src/yt_game_records.c
@@ -278,7 +278,7 @@ yt_planet_decode(struct yt_planet *planet, const struct yt_record *record)
 	planet->missiles = yt_record_get_number(record, YT_F69);
 	planet->owner = yt_record_get_number(record, YT_F73);
 	planet->ground_forces = yt_record_get_number(record, YT_F77);
-	planet->name_length = yt_record_get_number(record, YT_F85);
+	planet->name_length = (size_t)yt_record_get_number(record, YT_F85);
 	planet->last_minute = yt_record_get_number(record, YT_F89);
 	planet->plasma = yt_record_get_number(record, YT_F113);
 	planet->bank = yt_record_get_number(record, YT_F117);
@@ -310,7 +310,7 @@ yt_planet_encode(struct yt_planet *planet)
 	yt_record_set_number_if_changed(&planet->record, YT_F77,
 	    planet->ground_forces);
 	yt_record_set_number_if_changed(&planet->record, YT_F85,
-	    planet->name_length);
+	    (float)planet->name_length);
 	yt_record_set_number_if_changed(&planet->record, YT_F89,
 	    planet->last_minute);
 	yt_record_set_number_if_changed(&planet->record, YT_F113,
